% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.data.frame.qshap_result}
\alias{as.data.frame.qshap_result}
\title{Coercion method to data.frame for qshap_result}
\usage{
\method{as.data.frame}{qshap_result}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A qshap_result object}

\item{row.names}{Not used}

\item{optional}{Not used}

\item{...}{Additional arguments (currently unused)}
}
\description{
Coercion method to data.frame for qshap_result
}
\value{
A \code{data.frame} with columns \code{feature} (character) and
\code{rsq} (numeric), sorted by \code{rsq} in decreasing order.
}
