% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_explainer.R
\name{loss.f}
\alias{loss.f}
\title{Convenience wrapper for qshap_loss}
\usage{
loss.f(explainer, x, y, y_mean_ori = NULL)
}
\arguments{
\item{explainer}{A qshapr_tree_explainer object created by \code{gazer()}}

\item{x}{Feature matrix or data frame}

\item{y}{Response vector}

\item{y_mean_ori}{Optional pre-computed mean of y (for efficiency)}
}
\description{
A shorter alias for \code{\link{qshap_loss}}. See that function for full documentation.
}
\seealso{
\code{\link{qshap_loss}} for detailed documentation
}
