% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_qshapr_tree_explainer}
\alias{new_qshapr_tree_explainer}
\title{Constructor for qshapr_tree_explainer class}
\usage{
new_qshapr_tree_explainer(
  model,
  model_type,
  max_depth,
  base_score = NULL,
  trees,
  store_v_invc,
  store_z
)
}
\arguments{
\item{model}{The original tree model object}

\item{model_type}{Character string indicating model type ("xgboost" or "lightgbm")}

\item{max_depth}{Integer maximum tree depth}

\item{base_score}{Numeric base score (for XGBoost)}

\item{trees}{List of tree objects}

\item{store_v_invc}{Precomputed complex values for SHAP computation}

\item{store_z}{Precomputed root values for SHAP computation}
}
\value{
An object of class \code{qshapr_tree_explainer}
}
\description{
Creates a qshapr_tree_explainer object
}
\keyword{internal}
