% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qshap_result}
\alias{print.qshap_result}
\title{Print method for qshap_result}
\usage{
\method{print}{qshap_result}(x, n = 10, ...)
}
\arguments{
\item{x}{A qshap_result object}

\item{n}{Integer number of top features to display (default: 10)}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for qshap_result
}
