% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qshap_tree_explainer}
\alias{print.qshap_tree_explainer}
\title{Print method for qshap_tree_explainer}
\usage{
\method{print}{qshap_tree_explainer}(x, ...)
}
\arguments{
\item{x}{A qshap_tree_explainer object}

\item{...}{Additional arguments (currently unused)}
}
\value{
The input \code{x} is returned invisibly. Called primarily for its
  side effect of printing a summary of the \code{qshap_tree_explainer} object
  to the console.
}
\description{
Print method for qshap_tree_explainer
}
