% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qshapr_tree_explainer}
\alias{print.qshapr_tree_explainer}
\title{Print method for qshapr_tree_explainer}
\usage{
\method{print}{qshapr_tree_explainer}(x, ...)
}
\arguments{
\item{x}{A qshapr_tree_explainer object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for qshapr_tree_explainer
}
