% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.tree_summary}
\alias{print.tree_summary}
\title{Print method for tree_summary}
\usage{
\method{print}{tree_summary}(x, ...)
}
\arguments{
\item{x}{A tree_summary object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for tree_summary
}
\value{
The input \code{x} is returned invisibly. Called primarily for its side
effect of printing a summary of the \code{tree_summary} object to the
console.
}
