% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{qshap-classes}
\alias{qshap-classes}
\title{S3 Class Constructors and Methods for qshap}
\description{
This file contains formal S3 class definitions, constructors, validators,
and methods for the qshap package objects.
}
\keyword{internal}
