% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{qshapr-classes}
\alias{qshapr-classes}
\title{S3 Class Constructors and Methods for qshapr}
\description{
This file contains formal S3 class definitions, constructors, validators,
and methods for the qshapr package objects.
}
\keyword{internal}
