% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.qshap_result}
\alias{summary.qshap_result}
\title{Summary method for qshap_result}
\usage{
\method{summary}{qshap_result}(object, ...)
}
\arguments{
\item{object}{A qshap_result object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for qshap_result
}
