% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.qshap_tree_explainer}
\alias{summary.qshap_tree_explainer}
\title{Summary method for qshap_tree_explainer}
\usage{
\method{summary}{qshap_tree_explainer}(object, ...)
}
\arguments{
\item{object}{A qshap_tree_explainer object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Provides detailed summary information about the explainer
}
