% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_qshap_result}
\alias{validate_qshap_result}
\title{Validator for qshap_result}
\usage{
validate_qshap_result(x)
}
\arguments{
\item{x}{A qshap_result object}
}
\value{
The validated object (invisibly) or stops with an error
}
\description{
Validator for qshap_result
}
\keyword{internal}
