% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_qshap_tree_explainer}
\alias{validate_qshap_tree_explainer}
\title{Validator for qshap_tree_explainer}
\usage{
validate_qshap_tree_explainer(x)
}
\arguments{
\item{x}{A qshap_tree_explainer object}
}
\value{
The validated object (invisibly) or stops with an error
}
\description{
Validator for qshap_tree_explainer
}
\keyword{internal}
