% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_qshapr_tree_explainer}
\alias{validate_qshapr_tree_explainer}
\title{Validator for qshapr_tree_explainer}
\usage{
validate_qshapr_tree_explainer(x)
}
\arguments{
\item{x}{A qshapr_tree_explainer object}
}
\value{
The validated object (invisibly) or stops with an error
}
\description{
Validator for qshapr_tree_explainer
}
\keyword{internal}
