% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_simple_tree}
\alias{validate_simple_tree}
\title{Validator for simple_tree class}
\usage{
validate_simple_tree(x)
}
\arguments{
\item{x}{A simple_tree object}
}
\value{
The validated object (invisibly) or stops with an error
}
\description{
Validator for simple_tree class
}
\keyword{internal}
