% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_tree_summary}
\alias{validate_tree_summary}
\title{Validator for tree_summary class}
\usage{
validate_tree_summary(x)
}
\arguments{
\item{x}{A tree_summary object}
}
\value{
The validated object (invisibly) or stops with an error
}
\description{
Validator for tree_summary class
}
\keyword{internal}
