% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.simple_tree}
\alias{print.simple_tree}
\title{Print method for simple_tree}
\usage{
\method{print}{simple_tree}(x, ...)
}
\arguments{
\item{x}{A simple_tree object}

\item{...}{Additional arguments (currently unused)}
}
\value{
The input \code{x} is returned invisibly. Called primarily for its
  side effect of printing a summary of the \code{simple_tree} object to the
  console.
}
\description{
Print method for simple_tree
}
