% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.qshap_result}
\alias{summary.qshap_result}
\title{Summary method for qshap_result}
\usage{
\method{summary}{qshap_result}(object, ...)
}
\arguments{
\item{object}{A qshap_result object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for qshap_result
}
\value{
The input \code{object} is returned invisibly. Called primarily for its
side effect of printing a detailed summary of the \code{qshap_result}
object to the console.
}
